separate out light client and rpc server into packages
