+++ .gitignore
@@ -25,3 +25,4 @@ go.work.sum
 .env
 
 btclightclient
+!btclightclient/
