+++ btclightclient/btc_light_client_test.go
@@ -1,4 +1,4 @@
-package main
+package btclightclient
 
 import (
 	"errors"
