+++ btclightclient/lightblock.go
@@ -1,4 +1,4 @@
-package main
+package btclightclient
 
 import (
 	"math/big"
