+++ btclightclient/utxo_verification.go
@@ -1,4 +1,4 @@
-package main
+package btclightclient
 
 import (
 	"bytes"
