+++ btclightclient/utxo_verification_test.go
@@ -1,4 +1,4 @@
-package main
+package btclightclient
 
 import (
 	"encoding/hex"
